Order functionality
Able to add items to order
