+++ lib/cartScreen/cart-screen.dart
@@ -18,9 +18,9 @@ class CartScreen extends StatefulWidget {
 class _CartScreenState extends State<CartScreen> {
 
   final modifications1 = [
-    "Coke (regular)",
-    "Fries",
-    "Pickles (none)",
+    "Pepper Jack Cheese (Large)",
+    "Raspberry Vinaigrette",
+    "Pickles(none)",
     "Onions (extra)",
     "Ranch (side)"
   ];
@@ -188,7 +188,7 @@ class _CartScreenState extends State<CartScreen> {
                           children: <Widget>[
                             ElevatedButton(
                               onPressed: (){
-                                Navigator.push(context, MaterialPageRoute(builder: (context) => const ItemScreen(name: 'Double Burger Combo', img: 'assets/images/double-burger.jpg', price: 8.99, mainMods: ['Drink', 'Side'], dialogMods: ['BurgerMod', 'BurgerSauce'])));
+                                Navigator.push(context, MaterialPageRoute(builder: (context) => ItemScreen(name: 'Double Burger Combo', img: 'assets/images/double-burger.jpg', price: 8.99, mainMods: ['Drink', 'Side'], dialogMods: ['BurgerMod', 'BurgerSauce'])));
                               },
                               child: const Text('Edit', style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold, color: Colors.white)),
                               style: ElevatedButton.styleFrom(
